クラスを定義しているファイル名及び出力するファイル名の変更
