--- 04.class/main.js
@@ -1,12 +1,0 @@
-#!/usr/bin/env node
-
-import Memo from "./memo.js";
-import MemoDatabase from "./memo_database.js";
-
-(async () => {
-  const db = new MemoDatabase("memos.db");
-  await db.init();
-
-  const app = new Memo(db);
-  app.run();
-})();
